python/requirements.txt
black==22.3.0
click==8.0.4
dataclasses==0.8
flake8==4.0.1
importlib-metadata==4.2.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
piexif==1.1.3
Pillow==8.4.0
platformdirs==2.4.0
pycodestyle==2.8.0
pyflakes==2.4.0
tomli==1.2.3
typed-ast==1.5.2
typing_extensions==4.1.1
zipp==3.6.0
